requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib